auth_api/src/requirements.txt
Flask
PyJWT
mysql-connector